Added low-level IR compilation

+++ src/parser/tokenizer/token/inlined.rs
@@ -3,8 +3,8 @@ use std::fmt;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum Inlined {
-    Int(i32),
-    Float(f32),
+    Int(i64),
+    Float(f64),
     Bool(bool),
     Nothing,
 }
@@ -14,10 +14,10 @@ impl TryFrom<&[u8]> for Inlined {
 
     fn try_from(bytes: &[u8]) -> Result<Inlined, ()> {
         let as_str = String::from_utf8(bytes.to_vec()).unwrap();
-        if let Ok(int) = as_str.parse::<i32>() {
+        if let Ok(int) = as_str.parse::<i64>() {
             return Ok(Inlined::Int(int));
         };
-        if let Ok(float) = as_str.parse::<f32>() {
+        if let Ok(float) = as_str.parse::<f64>() {
             return Ok(Inlined::Float(float));
         };
         if let Ok(boolean) = as_str.parse::<bool>() {
